refactor: delay supported domains import
Import is really expensive cause it imports every crawler
