+++ cyberdrop_dl/ui/prompts/user_prompts.py
@@ -16,11 +16,6 @@ from rich.console import Console
 
 from cyberdrop_dl import __version__
 from cyberdrop_dl.constants import BROWSERS, RESERVED_CONFIG_NAMES
-from cyberdrop_dl.data_structures.supported_domains import (
-    SUPPORTED_FORUMS,
-    SUPPORTED_SITES_DOMAINS,
-    SUPPORTED_WEBSITES,
-)
 from cyberdrop_dl.ui.prompts import basic_prompts
 from cyberdrop_dl.ui.prompts.defaults import ALL_CHOICE, DONE_CHOICE, EXIT_CHOICE
 from cyberdrop_dl.utils.cookie_management import get_cookies_from_browsers
@@ -160,6 +155,8 @@ class DomainType(IntEnum):
 
 def domains_prompt(*, domain_message: str = "Select site(s):") -> tuple[list[str], list[str]]:
     """Asks the user to select website(s) for cookie actions and cache actions."""
+    from cyberdrop_dl.supported_domains import SUPPORTED_FORUMS, SUPPORTED_SITES_DOMAINS, SUPPORTED_WEBSITES
+
     OPTIONS = [["Forum", "File Host"], ["All Supported Websites"]]
     choices = basic_prompts.create_choices(OPTIONS)
     domain_type = basic_prompts.ask_choice(choices, message="Select category:")
@@ -200,6 +197,7 @@ def domains_prompt(*, domain_message: str = "Select site(s):") -> tuple[list[str
 
 def extract_cookies(manager: Manager, *, dry_run: bool = False) -> None:
     """Asks the user to select browser(s) and domains(s) to import cookies from."""
+    from cyberdrop_dl.supported_domains import SUPPORTED_FORUMS, SUPPORTED_SITES_DOMAINS, SUPPORTED_WEBSITES
 
     supported_forums, supported_websites = list(SUPPORTED_FORUMS.values()), list(SUPPORTED_WEBSITES.values())
     domains, all_domains = domains_prompt(domain_message="Select site(s) to import cookies from:")
